requirements.txt
langchain_community==0.4.1
langchain_core==1.3.2
langchain_groq==1.1.2
langgraph==1.1.9
python-dotenv==1.2.2
streamlit==1.56.0
supabase==2.29.0
tenacity==9.1.4
